Reduced amount of Direct Comparisons used in UV Tile Discarding
Originally, each UV Tile Checked and Discarded itself... which meant that per-vertex there were 32 if and && checks in the shader... that's really bad. Each check now Saturates and Multiplies the results and then returns that, everything is then combined and sent into a single if statement. Going from 32 per-vertex to 1 per-vertex.